requirements.txt
pandas==2.2.2
numpy==2.4.2
SQLAlchemy===2.0.32
psycopg[binary]>=3.1,<4.0
python-dotenv==1.0.1
tqdm==4.66.5